+++ CHANGELOG.md
@@ -54,3 +54,13 @@ All notable changes to this project will be documented in this file.
 
 - category icons
 - improved error feedback
+
+## 2025-08-09
+
+### Added
+
+- budget progress bars
+
+### Fixed
+
+- login/register display